refactor: indexer in light-client
