+++ prover/client/Cargo.toml
@@ -21,6 +21,7 @@ light-bounded-vec = { workspace = true }
 light-compressed-account = { workspace = true }
 light-batched-merkle-tree = { workspace = true }
 light-indexed-array = { workspace = true }
+
 ark-bn254 = { workspace = true }
 ark-serialize = { workspace = true }
 ark-std = { workspace = true, features = ["parallel"] }
